feat(forum_conversation): migrate CertifiedPost model from forum_upvote to forum_conversation
