+++ lacommunaute/forum_upvote/models.py
@@ -1,8 +1,7 @@
 from django.conf import settings
 from django.db import models
-from machina.models.abstract_models import DatedModel
 
-from lacommunaute.forum_conversation.models import Post, Topic
+from lacommunaute.forum_conversation.models import Post
 
 
 class UpVote(models.Model):
@@ -32,40 +31,3 @@ class UpVote(models.Model):
         ordering = [
             "-created_at",
         ]
-
-
-class CertifiedPost(DatedModel):
-    topic = models.OneToOneField(
-        Topic,
-        related_name="certified_post",
-        on_delete=models.CASCADE,
-        verbose_name="Topic",
-    )
-
-    post = models.OneToOneField(
-        Post,
-        related_name="certified_post",
-        on_delete=models.CASCADE,
-        verbose_name="Post",
-    )
-    user = models.ForeignKey(
-        settings.AUTH_USER_MODEL,
-        related_name="certified_posts",
-        blank=True,
-        null=True,
-        on_delete=models.SET_NULL,
-        verbose_name="User",
-    )
-
-    class Meta:
-        ordering = [
-            "-created",
-        ]
-
-    def __str__(self):
-        return f"{self.post} - {self.topic} - {self.user}"
-
-    def save(self, *args, **kwargs):
-        if self.topic != self.post.topic:
-            raise ValueError("The post is not link to the topic")
-        super().save(*args, **kwargs)
